requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
requests==2.32.3
python-dotenv